+++ rust/src/config.rs
@@ -37,6 +37,7 @@ pub struct ToolchainTweaks {
     pub treat_j_as_unconditional_branch: bool,
     pub sn64_div_fix: bool,
     pub gnu_mode: bool,
+    pub r5900_prodg_sn_as_inverted_regs: bool,
 }
 
 #[repr(C)]