+++ rust/src/instruction.rs
@@ -483,6 +483,13 @@ impl Instruction {
         self.flags = utils::bitrepack(self.flags, value.into(), 2, 2);
     }
 
+    pub fn flags_get_r5900_prodg_sn_as_inverted_regs(&self) -> utils::TrinaryValue {
+        utils::shiftr(self.flags, 4, 2).try_into().unwrap()
+    }
+    pub fn flags_set_r5900_prodg_sn_as_inverted_regs(&mut self, value: utils::TrinaryValue) {
+        self.flags = utils::bitrepack(self.flags, value.into(), 4, 2);
+    }
+
     pub fn instr_id_type_name(&self) -> &'static str {
         unsafe { std::ffi::CStr::from_ptr(RabInstrIdType_getName(self.instr_id_type)) }
             .to_str()